sink(ticdc): split config test
